html/browsers/browsing-the-web/navigating-across-documents/javascript-url-rules.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol">
<link rel="help" href="https://html.spec.whatwg.org/multipage/webappapis.html#run-a-classic-script">
<link rel="help" href="https://url.spec.whatwg.org/#percent-decode">
<link rel="help" href="https://encoding.spec.whatwg.org/#utf-8-decode">
<title>javascript: url processing rules</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>

<!-- 4, %0a is treated as a byte and turned into "\n" (0x0A is \n), which means "b" is the return value -->
<iframe id="a" src='javascript:"b";%0a"a";'></iframe>
<!-- 4, %25 is treated as a byte and turned into "%" (0x25 is %) -->
<iframe id="b" src='javascript:"%2525%25"'></iframe>
<!-- 9, Script that will be unsuccessful, result is undefined -->
<iframe id="c" src='javascript:"use strict"; foo = 1;'></iframe>
<!-- 10, evaluation was unsuccessful, so result is undefined -->
<iframe id="d" src='javascript:throw new Error(Number.MAX_SAFE_INTEGER)'></iframe>
<!-- 9 & 4, %24 is treated as a byte and turned into "$" (0x24 is $) -->
<iframe id="e" src='javascript:var $a = Number.MAX_SAFE_INTEGER; `${%24a}`;'></iframe>

<script>
async_test(function(test) {
  function check(id, expected) {
    assert_equals(
      document.getElementById(id).contentDocument.body.textContent,
      expected
    );
  }
  window.onload = test.step_func_done(function() {
    check("a", "a");
    check("b", "%25%");
    check("c", "");
    check("d", "");
    check("e", String(Number.MAX_SAFE_INTEGER));
  });
}, "'javascript: url' contents processing rules");
</script>
